backend/requirements.txt
Django>=5.1,<5.2
django-ninja>=1.3
django-cors-headers>=4.6
python-decouple>=3.8
argon2-cffi>=23.1
PyJWT>=2.8
